findbugs says...

+++ examples/gsod/src/main/java/org/opennms/newts/gsod/ExportRunner.java
@@ -76,7 +76,7 @@ public class ExportRunner {
         System.out.printf("timestamp,%s%n", m_metric);
 
         for (Results.Row<Sample> row : m_repository.select(m_resource, timestamp(m_start), timestamp(m_end))) {
-            System.out.printf("%s,%.2f%n", row.getTimestamp().asDate(), row.getElement(m_metric).getValue());
+            System.out.printf("%s,%.2f%n", row.getTimestamp().asDate(), row.getElement(m_metric).getValue().doubleValue());
         }
 
         return 0;
